Add Adam Barrow's Battery Voltage patch
This adds the beginnings of a framework for parsing incoming commands on the serial port.  It's currently only used for battery voltage messages from the optional oil pan board.  We'll probably have to change the format as we expand this later, but it's nice to get the ball rolling on serial input.

This also reorders functions in serialIO.c to hopefully make a bit more sense.


git-svn-id: https://gentlenav.googlecode.com/svn/trunk@245 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8