Updated flan.pyw with Peter Hollands' help. This version of flight analyzer will attempt to use GCC if the Windows "program files" environment variable is not available. Tested on Mac OS X.6 with gcc 4.2.1 and Ubuntu with gcc 4.1.1. The assumption is that you will have /bin/sh installed and gcc will be configured in your path.
Also corrected an issue when running without the 'Models' directory where the error message was displayed briefly onscreen but not really observable by the user. It now displays a popup message in this case, requiring acknowledgement from the user.

Please note that this version still does not handle waypoints with arithmetic (CORNER + 75). We'll probably work on that next :).

Changes to functions:
* Tools/flight_analyzer/flan.pyw
  (C_pre_processor): 
  (convert_to_absolute_lat_long): 
  (wrap_kml_into_kmz): 



git-svn-id: https://gentlenav.googlecode.com/svn/trunk@407 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8