Fix a bug that caused a brief high-throttle when entering RTL mode
When the radio signal is first lost, radio_on is set to 0, but the state machine is not run immediately.  This delay caused a short period where radio_on is 0, but the plane is not yet in RTL mode, so the waypoint flags (including F_LAND) were not yet set up correctly.  This could lead to inadvertent throttling up until the state machine finally ran, up to 1/2 second later.

Solved this by adding a callback for loss of radio signal, and using it to set up the flags correctly immediately upon radio loss.


git-svn-id: https://gentlenav.googlecode.com/svn/trunk@934 cf4e4d4a-e611-11dd-8fb1-9960f2a117f8