Before the number of instructions in interrupts should not exceed MAX_INSTRUCTIONS_PER_CYCLE. Simply making MAX_INSTRUCTIONS_PER_CYCLE larger would cause 40(Hz) * x instructions per seconds
This fix: the first cycle will run the first MAX_INSTRUCTIONS_PER_CYCLE instructions,
but will continue with the next cycle to run the instructions that follow the first batch. If no CLEAR is encountered, the routine will run to its end and start over.
Interrupt routines are only run while the main program is waiting for a fly command to finish.